requirements.txt
# --- 基础工具 ---
python-dotenv>=1.0.0
httpx>=0.27.0
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0

# --- MCP 协议 ---
mcp[cli]>=1.0.0

# --- Agent 框架 ---
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# --- Web & Audio ---
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
jinja2>=3.1.0
# 豆包文档主要使用 websocket 通信，PyAudio用于采集(Mac需先安装portaudio)
websockets>=12.0
# pyaudio>=0.2.14